add convertStatisticsToHistory utility
